small refactor changes
